Create README.md
You are asked to train a feed forward network to recognize handwritten digits. Use the MNIST data commonly available in the internet.
1. Read your data.
2. Start with AlexNet architecture. Change the input layer to handle grayscale images. Change the number of outputs to 10 each representing one digit.
3. Train the network with the data and report results.
a. Use at least two different learning rate adjustment schemes.
b. Use at least three different activation functions.
4. Change the network architecture.
a. Change the number of nodes in the fully connected layer by 10% for three times and repeat Step 3.
b. Continuing with %20 of the nodes in the fully connected layer, remove the third layer from the output and repeat Step 3.
c. Continuing with %10 of the nodes in the fully connected layer, remove the third and fourth layers from the output and repeat Step 3.